Phase 8: fix Linux/macOS clippy regressions surfaced by CI
Two clippy errors that only fire on Linux/macOS targets — the Windows
maintainer host can't catch these because the offending code lives
inside a non-Windows cfg branch and is dead-stripped before clippy
sees it.

- daemon/mythd-windows/src/wsl_bridge.rs: `use std::process::Command`
  was a top-level import but `Command` is only used inside the
  `#[cfg(target_os = "windows")]` `list_distros()` branch. On
  Linux/macOS clippy that branch is excluded and the import was
  unused. Fix: drop the top-level import and inline
  `std::process::Command::new(...)` at the call site.

- daemon/mythd-linux/src/fanotify.rs: the `fanotify_init` flags cast
  was `(FAN_CLASS_CONTENT | FAN_CLOEXEC) as u32`. On Linux those
  constants are already `c_uint` (u32), so the cast was a
  same-type cast clippy rejects under -D warnings. The Windows
  maintainer's local cargo build never compiles the Linux body
  of `FanotifyHandle::open` so this slipped through. Fix: drop
  the redundant cast; keep `O_RDONLY as u32` (genuine i32 -> u32).

Both fixes verified locally on Windows clippy + tests stay green.

+++ daemon/mythd-linux/src/fanotify.rs
@@ -102,8 +102,11 @@ impl FanotifyHandle {
         // below; further hardening (FD_CLOEXEC race etc.) is
         // captured in TASK-073's review pass.
         let fd = unsafe {
+            // FAN_CLASS_CONTENT | FAN_CLOEXEC are already `c_uint` on
+            // Linux per the libc crate — no cast needed. O_RDONLY is
+            // `c_int` and DOES need the cast.
             libc::fanotify_init(
-                (libc::FAN_CLASS_CONTENT | libc::FAN_CLOEXEC) as u32,
+                libc::FAN_CLASS_CONTENT | libc::FAN_CLOEXEC,
                 libc::O_RDONLY as u32,
             )
         };
